# src/main/resources/static/swagger-apis/api1/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a channel which connects U to all the bank services :)",
    "version": "1.0.0",
    "title": "Idea Channel Application"
  },
  "host": "localhost:9000",
  "tags": [
    {
      "name": "deposits",
      "description": "Everything about your Deposits"
    },
    {
      "name": "General-APIs",
      "description": "General Controller"
    }
  ],
  "paths": {
    "/idea-channel/banks/branches": {
      "post": {
        "tags": [
          "General-APIs"
        ],
        "summary": "getBankBranchList",
        "operationId": "getBankBranchListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "body",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/idea-channel/deposits/balance": {
      "post": {
        "tags": [
          "deposits"
        ],
        "summary": "get deposit balance by deposit number",
        "operationId": "getdepositsBalanceUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "get deposit info",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "headers",
            "in": "header",
            "description": "headers",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/idea-channel/deposits/info": {
      "post": {
        "tags": [
          "deposits"
        ],
        "summary": "getdepositsInfo",
        "operationId": "getdepositsInfoUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "body",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "headers",
            "in": "header",
            "description": "headers",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/idea-channel/deposits/statements": {
      "post": {
        "tags": [
          "deposits"
        ],
        "summary": "getdepositsStatements",
        "operationId": "getdepositsStatementsUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "body",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "headers",
            "in": "header",
            "description": "headers",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  }
}